BlackSkullOfMastery for Necromancer, TomeOfMastery is for the rest

+++ PixelDungeon/src/main/java/com/watabou/pixeldungeon/actors/mobs/Tengu.java
@@ -28,6 +28,7 @@ import com.watabou.pixeldungeon.actors.Actor;
 import com.watabou.pixeldungeon.actors.Char;
 import com.watabou.pixeldungeon.actors.blobs.ToxicGas;
 import com.watabou.pixeldungeon.actors.buffs.Poison;
+import com.watabou.pixeldungeon.actors.hero.HeroClass;
 import com.watabou.pixeldungeon.effects.CellEmitter;
 import com.watabou.pixeldungeon.effects.Speck;
 import com.watabou.pixeldungeon.items.TomeOfMastery;
@@ -74,7 +75,10 @@ public class Tengu extends Boss {
 	
 	@Override
 	public void die( Object cause ) {
-		Dungeon.level.drop( new TomeOfMastery(), getPos() ).sprite.drop();
+
+		if ( Dungeon.hero.heroClass != HeroClass.NECROMANCER){
+			Dungeon.level.drop( new TomeOfMastery(), getPos() ).sprite.drop();
+		}
 
 		GameScene.bossSlain();
 		Dungeon.level.drop( new SkeletonKey(), getPos() ).sprite.drop();
